rdp: Run do_render() as GSourceFunc in custom GMainContext
In order to be able to run all graphics operations in a different
thread, add a new GSource, which runs do_render() as callback function.
This GSource will then be attached to a custom GMainContext.
In a future commit, this custom GMainContext will be the GMainContext
of the graphics thread.