session-rdp: Handle all graphics operations in a separate thread
Currently, all graphics operations happen in the main thread, which
means that remote input is delayed until the current graphics
operation, such as encoding frame content for example, is done.
On very slow servers or with the raw bitmaps path in the legacy path,
this highly reduces the input responsiveness, as the main thread can
become fully occupied.

So, run all graphics operations in a separate thread, while the main
thread takes care of pushing the input events to mutter.