daemon: Exit, if all backends fail to initialize
When gnome-remote-desktop is started, it attempts to initialize the RDP
and VNC backend, depending on with which backends gnome-remote-desktop
was built.
However, if all backends fail to initialize, gnome-remote-desktop will
still continue to run, despite not being usable.
This is especially problematic, as the systemd user service still runs
as if there would be no problem, making it hard to find the cause why a
connection to the server won't make it.

To get rid of this situation, also check the result of the backend
initialization.
If all backends fail, exit the daemon, as there is no reason to
continue running.