rdp: Get rid of the grd_session_rdp_resize_framebuffer() API
Currently, when the PipeWire stream parameters of the monitor content
change, grd_session_rdp_resize_framebuffer() is called.
However, this call is not necessary and can be delayed right before the
next graphics update.
In addition to that, this also allows to easily run do_render() later
in a different thread, as no order of graphics output buffer resizes
and frame updates needs to be considered.

So, remove the grd_session_rdp_resize_framebuffer() API from the
session-rdp header file and make this function private.
Call this private function then at the beginning of
grd_session_rdp_take_buffer() to ensure that the graphics output buffer
is correctly resized.