rdp-fuse-clipboard: Always explicitly use global main context
While g_main_context_get_thread_default() refers in the context of the
timeout_reset_source creation to the global main context, this function
can technically also return a different main context.
The intention is, however, that the timeout_reset_source always runs on
the main thread.

Thus, change the main context, to attach the GSource to, to NULL, which
refers to the global main context.