session-rdp: Default to max 60 FPS, when using graphics pipeline
With the graphics pipeline, gnome-remote-desktop has a way to control
to control the encoding rate based on the speed of decoding and
displaying frames on the client side.
With the PipeWire stream running at 60 FPS also decreases the frame
latency between when the frame was rendered by mutter and when
gnome-remote-desktop can start encoding it.
A 60 FPS PipeWire stream does not mean that 60 are always reached.
The maximum FPS is determined by the minimum of the encoding rate in
gnome-remote-desktop and the frame acknowledge rate of the RDP client.