session-rdp: Only signal GCond, when all encode jobs are complete
Currently, when using the raw bitmaps path, every worker thread in the
raw bitmaps path signals the GCond for the thread pool once their task
is done.
However, this is wasteful, since the only necessary signal here is the
final one, where the last tile is being encoded.

So, make the signal call conditional. For a FullHD frame this saves
about 509 unnecessary wakeups of the graphics thread.