rdp-fuse-clipboard: Fix potential race when accessing FuseFile path
When outputting debug information about which FuseFile is opened, the
FuseFile path might already be invalid, since the FuseFile could
already be deleted due to a new clipboard selection.
While this situation is unlikely to happen, since pasting content and
copying content would have to happen at the same time, it is in theory
still possible.

So, duplicate the FuseFile path before unlocking the filesystem mutex.
This will avoid this situation.