rdp/rdpgfx: Unref surface serial after looking up surface context
The surface context of a surface exists until the surface is deleted
and all references are cleared.

So, when looking up a the surface context to ack a frame, unref the
serial afterwards and not before, as otherwise gnome-remote-desktop
would crash since the surface context might already be deleted, due to
the already been deleted.
Currently, gnome-remote-desktop is not able to crash with its current
architecture here, since only one monitor is screencasted and the GFX
surface currently exists until the end of the remote desktop session.