clipboard: Split up SelectionTransfer handling
Currently, when gnome-remote-desktop handles a SelectionTransfer-
signal, the mime type content is requested synchronously from the
respective backend.
Since this process is done synchronously, the main thread is blocked
and cannot handle any other events, such as input events.
With additional network latency, this becomes worse.

To allow the backend implementations to transform to an asynchronous
implementation, split up the SelectionTransfer handling.
The signal will now just emit the request on the backend and the
backend will, when it has the response ready submit the response
itself.

Both the RDP and VNC backend will currently still function as before,
as the handling is just split up.
Since the VNC backend does not support delayed rendering of clipboard
data, no further changes will happen for the VNC backend.
The RDP backend, on the other hand, will be reworked in the next commit
to handle mime type content requests asynchronously.