session: Only handle Selection* signals when clipboard is enabled
When the RDP or VNC backend disable the clipboard, since the session is
about to be destroyed, then gnome-remote-desktop could still receive
SelectionOwnerChanged- and SelectionTransfer-signals.
Currently, these signals are handled unconditionally.

To ensure that this won't be the case any more, ignore these signals,
when the clipboard instance is unavailable.