rdp/nw-auto: Only create ping source if required and non-existent
When updating the ping source, gnome-remote-desktop clears the previous
GSource first, if it is required.
If it is not required, then gnome-remote-desktop won't clear the
existing GSource.
This behaviour is correct. However, when it is not required to update
the ping source, gnome-remote-desktop will still attempt to create a
new GSource and replace the existing GSource with the new one.
The old GSource will then remain to exist and cannot be removed any
more, since no reference to that GSource is left.

Fix this behaviour by just returning from the function in such
situation.
Currently, gnome-remote-desktop does not run into this situation, since
it seems that all known RDP clients seem to behave well with the
SuppressOutput PDU.