settings: Add new screenshare mode setting for the RDP backend
Implementing support for virtual monitors should be done without
regressions.
By default, gnome-remote-desktop currently mirrors the primary monitor.
For now, this behaviour should be preserved.
In order to be able to use virtual monitor in a remote desktop session,
a switch is needed.

So, add a new gsetting, which allows choosing between remote desktop
sessions, where the primary monitor is recorded and remote desktop
sessions, which create and use a virtual monitor.
The latter also allows the usage of gnome-remote-desktop, when no
physical monitors are attached.