session: Stop trying to stop the session, when it is already destroyed
In order to start a remote desktop session, gnome-remote-desktop calls
several dbus methods on the remote desktop- and screencast interface of
mutter.
These calls use in glib internally GTasks and will always run their
async-ready-callback, when the task is done, aborted, or cancelled.
This also applies to situations, where the session is immediately
stopped or destroyed.
In such cases, the callback should not be handled any more, as parts of
the session are already destroyed, due to the session already been
stopped, or the complete session is already destroyed.
The result is usually a crash, as gnome-remote-desktop tries to stop
the session, although it is already stopped or destroyed.

To prevent these situations, always cancel the cancellable, when
stopping the session.
When a async-ready-callback is called, don't directly derenference the
user date.
Instead, call the finish function first and check the error code.
When the cancellable was cancelled, the finish function will not return
successfully and set the error.
The error code will in such cases be G_IO_ERROR_CANCELLED.
If that is the case, just return, as the session is already stopped.