settings: Set error if username or password is NULL
When looking up the username or password for an RDP session,
gnome-remote-desktop will parse the credentials string, returned via
libsecret.
This credentials string may exist, but it can contain garbage values,
when e.g. the user changed them to such garbage values.
In such case, the parsed username or password may be NULL, leading into
gnome-remote-desktop to crash, as the error is not set.

Fix this behaviour by always setting the GError, when the username or
password is NULL.