session: Require stream for absolute pointer motion events
When dealing with multi-monitor setups in the future, and
gnome-remote-desktop gets an absolute pointer motion event, it needs to
somehow identify the stream.
The session class won't be able to choose the correct stream in
multi-monitor setups, the respective backend (RDP or VNC), however,
can.

So, require the stream instance to be submitted for each absolute
pointer motion event.