rdp: Track and use rdp_surface reference in stream class
An rdp-pipewire-stream will always correspond to an RDP surface.
To be able to track all frame buffers of the pipewire-stream later,
track a reference to the RDP surface.
This will be relevant in the future, when resizing pipewire streams
during a session, which is usually the case with virtual monitors.