ci: Build mutter from git to have the updated screen cast API
mutter on current main has support for adding screencast sessions, when
the remote desktop session is already active.
This allows gnome-remote-desktop to perform monitor configuration
changes during an active remote desktop session.
Although the current remote desktop and screencast interface lacks some
features for multi monitor support, like having a way to rearrange the
monitors, it is still desired to have the new API, as it allows
gnome-remote-desktop to have similar paths for remote assistance and
headless sessions in the future.