rdp: Add RDP monitor config
RDP provides three different ways for clients to submit monitor
configurations:

1. via the Client Core Data (only single-monitor setups possible)
2. via the Client Monitor Data (allows multiple monitors)
3. via the Display Control Channel Extension

The first two ways are only used for the initial monitor configuration,
when connecting to an RDP server.
The third way is used to update the monitor configuration, when the
remote desktop session is already established.

To be able to handle monitor configurations easily, add a new helper
struct with helper functions, which allows to create internal monitor
configurations from any of these three sources.