session-rdp: Add support for sessions with virtual monitors
To implement support for sessions with virtual monitors, use the RDP
monitor configuration helper structs and functions to get the client
sided monitor configuration.
Since for multi-monitor support additional APIs or PipeWire stream
properties are required, limit the maximum amount of virtual monitors
to 1.
Use the internal monitor configuration to set the initial monitor
resolution of the virtual monitor.
To do this, pass the virtual monitor to the PipeWire stream instance.

Also, always allow input in sessions with virtual monitors, as these
sessions are not remote assistance sessions.