hwaccel-nvidia: Add CUDA kernels for damage region detection
As already mentioned in previous commits, damage region detection on
the GPU is much faster than damage region detection on the CPU.
Although the CPU has a higher clock rate, the GPU has many more cores,
which makes it more suitable for parallel tasks.

The damage detection on the GPU is done in four parts.

Part 1, the actual damage detection:
The check_damaged_pixel cuda kernel checks for each pixel, whether it
is damaged or not.
To make the implementation easier and to avoid accessing global memory
too much, read the values as 32-bit unsigned integers.
Before this kernel is run, the global region_is_damaged value is set to
0.
After running this kernel, gnome-remote-desktop knows whether the
region is damaged or not.
Since, NVENC does not make use of damage regions, only the
region_is_damaged value is retrieved, when using NVENC.
The rest of the region will still be retrieved, as clients like
xfreerdp still use it for optimization purposes.
This retrieval will run later, when NVENC encodes the data, as CUDA
kernels and data transfer between the GPU and CPU can still run, while
NVENC is encoding, thus saving time.

Part 2, Combining the coloums:
As gnome-remote-desktop uses damage tiles of the size of 64x64,
especially for RFX (Progressive), the damage values need to be
combined.
For this, the combine_damage_array_cols kernel runs multiple times.
First, all neighbours are combined, i.e. damage values of pixels 0 and
1 are combined in 0, and damage values of pixels 2 and 3 are combined
in pixel 2.
In the next round, the combine_shift is increased, leading to the
damage values of pixels 0 and 2 being combined in 0 and the damage
values of pixels 4 and 6 being combined in 4.
The same thing happens to the other pixels.
The combine_damage_array_cols kernel runs 6 times, while increasing the
combine_shift value in each run.
After this, every 64th value combines all damage values for maximum 64
pixels in each row.

Part 3, Combining the rows:
This is pretty much the same as part 2, except that this kernel now
runs on the rows, instead of the coloums.

Part 4: Simplifying the damage array:
With now every 64x64th value combines all damage values for its tile,
simplify the resulting damage array.
In case of the data needs to be transferred to the CPU, this
drastically increases the speed, as the simplified damage array is very
small.

These CUDA kernels will later be used by the CUDA damage detection
class.