egl-thread: Add fallback with EGL_PLATFORM_DEVICE_EXT
When initializing the EGL thread with the platform
EGL_PLATFORM_SURFACELESS_MESA, it won't be hardware accelerated, when
using the NVIDIA drivers.
The result is that EGL Thread will be torn down, although it could
actually be used with the NVIDIA drivers, even when dma-buf are not
used:

With CUDA, it is possible to map pixel buffer objects for the use in
CUDA.
These mapped CUDA buffers can then be used for e.g. the damage region
creation on the GPU, which is much faster than the creation on the CPU,
or be used in NVENC later.
The EGL thread can then, when dma-bufs are used, directly map the frame
data.
When dma-bufs are not available, the frame data can be uploaded, which
is done in future commits.
With this operation in the EGL thread, it won't block the graphics
thread in the RDP backend, which allows the RDP backend to maximize the
maximum possible framerate, due to shorter encoding times, as there
won't be a blocking upload operation be any more.

It would also make the implementation with the support of dma-bufs
easier in the future.
With dma-bufs, the frame data can then directly mapped, while being
implementation-wise close to the non-dma-buf path.

So, add a fallback path using the platform EGL_PLATFORM_DEVICE_EXT to
enable the ability to directly map pixel buffer objects, when using the
NVIDIA driver.