hwaccel-nvidia: Use compute devices retrieved from EGL thread
As the EGL thread will later upload data to the GPU or bind EGL images
from dma-bufs for the use in CUDA, the CUDA device should be the same
one as in the EGL thread.
To ensure this, CUDA provides an API to retrieve the current CUDA
devices associated to the current GL thread.

So, use this API to retrieve the CUDA devices from the EGL thread and
use them for CUDA and NVENC.
Also use the helper struct and functions from the last commit for the
signal-wait mechanism, as the class for hardware acceleration for
NVIDIA devices is created in the main thread.