Add grd-utils
There are several situations in gnome-remote-desktop, where one thread
has to wait on another thread to finish a certain task.
For these situations, a combination of a GCond, a GMutex, and a boolean
value is used.
This signal-wait mechanism has always the same pattern.
In order to avoid implementing it over and over again, create a helper
struct with helper functions, which take over the mechanism.