rdp: Add RDP buffer pool
Allocating buffers on the GPU or registering pixel buffer objects for
the use on the GPU takes more time and requires more effort than
allocating a buffer in memory.
Unregistering such buffers can introduce additional complexity to this
problem.

To take care of this problem, introduce a buffer recycle mechanism in
form a buffer pool.
This buffer pool allocates N buffers by default.
Instead of allocating a new buffer, when one is needed, the entity that
needs a new buffer takes a buffer from the pool.
If all buffers of a buffer pool are used, the buffer pool takes care of
allocating another buffer in order to handle such situations
gracefully.
Instead of deleting a buffer, when it is not needed any more, it will
be returned to the buffer pool.
Returning a buffer is a fast action. When a buffer pool resize is
needed upon returning a buffer, the buffer pool queues a resize using a
GSource.
This GSource will then resize the buffer pool on the main thread.