build: Stop making the nv-codec-headers optional
The main reason to make the nv-codec-headers optional in GNOME 41 was
that Debian and Ubuntu did not have an updated version of it and it
wasn't clear, whether both will update their version.

As it turns out, Debian (and therefore Ubuntu too) did update their
packages and also enabled hardware acceleration using CUDA and NVENC in
gnome-remote-desktop 41.
Since these headers are MIT licensed and don't provide any patented
functionality, as they only load functions from the NVIDIA driver, if
it is available, they can be made obligatory.

This will also make reading the code of the framebuffer handling much
easier in the future, when implementing the mechanism to handle all
frame data on the GPU.

Therefore, remove the NVENC/CUDA option and make the nv-codec-headers
obligatory for the RDP backend.