rdp: Allocate and register GL resources for each RDP buffer
When uploading or mapping frame data on the GPU later, the buffer for
the frame on the GPU is usually supposed to be already allocated and
registered at the external API, which makes use of the mapped buffer.

To ensure this, allocate and register the GL resources, when the size
of the framebuffers in the pipewire stream is known.
When the pipewire stream is resized, these framebuffers are resized
too.
This is relatively easy, as all framebuffers are tracked.

When a new buffer is added to the buffer pool, due to all buffers of
the buffer pool already being used, do not allocate any GL resources.

This has two reasons:
First, the EGL thread is about to take care of the framebuffer anyway,
so adding additional synchronization for the allocation can impact the
performance.
The additional synchronization is still required, as the allocation can
fail, which is a situation, that needs to be handled.

Second, the submission of the allocation task and synchronization would
happen on the PipeWire realtime thread, instead of the main thread,
which can lead to a deadlock, where the buffer pool mutex won't be
unlocked any more, since the EGL thread would wait on the PipeWire
realtime thread to unlock it and the PipeWire realtime thread would
wait on the EGL thread to finish the allocation task.