hwaccel-nvidia: Load extra CUDA functions for GL PBO mappings
The nv-codec-headers don't load all functions from the NVIDIA driver.
Two of these functions are cuGraphicsGLRegisterBuffer() and
cuGraphicsResourceGetMappedPointer().
These two functions are required in order to be able to map pixel
buffer objects in GL.
As there is no obvious way to contribute to the nv-codec-headers and
having a new version ready in Fedora for the CI image takes a lot of
time, load these functions in gnome-remote-desktop manually.
Since dlopen() is refcounted, this won't cause any problems.

So, load these two extra functions in order to be able to map pixel
buffer objects.