rdp: Add classes for damage region detection
Currently, the damage detection in gnome-remote-desktop is done using
memcmp.
When a frame is not completely different than its previous frame, which
is usually the case, then the damage detection using memcmp can take
several milliseconds.
Comparing frame buffers on the GPU is much faster, even when an upload
of the frame data to the GPU is required, but this method is not always
available.

To take care of this problem, add a new abstract class, which provides
function pointers of an implementation.
Also provide one fallback implementation using memcmp.

This allows, in the future, the implementation of other region
detection classes without requiring additional changes in the graphics
thread.