rdp: Add CUDA damage detector class
With the CUDA kernels in place, add now the damage detector class.
All tasks happen on the GPU happen asynchronously except for the
methods is_region_damaged() and get_damage_region(), as in both cases
the results need to be retrieved from the GPU.

Comparing the whole framebuffer itself with CUDA takes about 230µs on a
GTX660 for a FullHD frame, while memcmp takes for a whole framebuffer
about 6ms.
Combining the results of the damaged pixels takes around 500µs.

With these values, the damage detection with CUDA is in almost all
cases faster than memcmp.
The only case, where memcmp is fast, is when all pixels in the current
framebuffer are different, than in the previous framebuffer.
However, the difference is just about 100µs.