rdp-buffer-pool: Prefer buffers without mapped resources
When taking a buffer out of the buffer pool, the GL resources might
still be mapped from its last use.
While unmapping the pixel buffer object is a matter of tens of
microseconds, it still adds a delay.
Furthermore, when no frames are retrieved in gnome-remote-desktop, due
to the system being idle, the GPU driver downclocks the GPU.
The result is that the first two unmapping operations take several
milliseconds to complete, when gnome-remote-desktop receives some
frames again.

This, however, does not apply to the upload operation for some reason.
In addition to that, the mapping operations seem to perform faster for
some reason.
It might be the case, that the GPU driver already starts upclocking the
GPU, when data is uploaded to the GPU.

With this in mind, always prefer buffers, that don't need an unmap
operation.
In the next commits, the buffer pool receives a low priority GSource,
which unmaps all untaken buffers, when the main thread idles.