rdp-buffer: Add API to unmap resources
This API will be used in the next commit by the buffer pool to unmap
the GL resources of the buffer.

To unmap the GL resources, use the custom task API of the EGL thread to
queue an unmap.
Since all tasks in the EGL thread run in a queue, no synchronization is
required here.