rdp: Use internal enum for RDP errors
When performing damage detection on the GPU in the future, some API
will be used.
For NVIDIA GPUs, this will be CUDA. For other GPUs, similar APIs will
be used.
Although these API calls usually never fail, there is no guarantee,
that this will always be the case.
Such cases are rare, but need to be handled. The easiest thing to do is
to simply end the session.

Currently, this is done using the internal notify_error() API in
session-rdp.
However, this requires the knowledge of the FreeRDP error code and
including FreeRDP headers in certain classes, but it can lead to
compilation errors due to name collisions.

To take care of this problem, use an internal enum for RDP errors and
translate these errors into RDP errors in session-rdp.