rdp/rdpgfx: Forbid protocol resets without AVC support by the client
Clients, that advertise capability sets with version
RDPGFX_CAPVERSION_103 or later, can reset the graphics pipeline any
time.
This behaviour is only used by mstsc and also then only in some error
cases.

Since a reset of the graphics pipeline can imply, that the client won't
advertise support for AVC any more, gnome-remote-desktop always needs
to have a local copy of the frame data on the host, as encoding with
the RFX Progressive codec can currently only be done with the CPU and
with a protocol reset, the frame data might actually need to be
downloaded to the host, which would add huge amount of additional
complexity of the code paths.
Since mstsc will not stop advertising support for AVC, a reset of the
graphics pipeline without AVC support is unlikely to happen, but in
theory it is possible, as other clients can theoretically make use of
this protocol reset too.

To be able to use full hardware acceleration, without any transfer of
raw frame data between the Host and the GPU, protocol resets without
AVC support on the client side need to be forbidden.

So, do exactly that, to allow gnome-remote-desktop to avoid unnecessary
frame data transfer between the Host and the GPU.