rdp: Split up AVC420 encoding
Calling nvEncEncodePicture() does not block until the encoding
operation is done.
Locking the bitstream, on the other hand, is a blocking operation.

This allows doing other tasks before calling nvEncLockBitstream() to
lock the bitstream, which will be done in the next commit.

So, split up the AVC420 encoding to create the opportunity for this
optimization.