rdp: Defer cairo region creation if possible
NVENC does not make use of damage regions, but xfreerdp uses damage
regions to cut out unnecessary operations to display the decoded frame.

When NVENC encodes a frame, CUDA kernels can still run and data can
still be transferred between the Host and the GPU, without affecting
the NVENC operation.
With the split up of the NVENC encode operation done in the last
commit, run the cairo damage region creation, while NVENC encodes the
frame.
As the damage checking is already done, the damage region just needs to
be simplified and transferred to the host.
When this task is done, the graphics thread will wait for NVENC to
finish the encoding operation.