egl-thread: Also allow bound dma-bufs to be mapped
Currently, hardware acceleration using CUDA can only be used, when
dma-buf support is disabled, as mapping pixel buffer objects created
from dma-buf textures is not implemented yet.
In order to be able to make use of damage detection on the GPU, the EGL
thread needs to map the pixel buffer object for a dma-buf texture.
When NVENC support is unavailable, then the dma-buf framebuffer also
needs to be downloaded.
In the case of CUDA is unavailable, but dma-bufs are still supported,
gnome-remote-desktop needs to download the dma-buf framebuffer.

To support all these cases with dma-bufs, extend the current download
API accordingly.
When the dst_data is NULL, don't download the framebuffer from the GPU.
When a GrdEglThreadImportIface instance was provided to the download
function, import the dma-buf with the help of the provided functions in
the iface.
This allows later to directly use the dma-buf content for the use in an
external API like CUDA without additional GPU <-> CPU transfer.