rdp: Avoid local copy if possible
If the GPU supports NVENC, the client did not disable AVC, and an NVENC
session was successfully opened, there is no need to have the frame
data available to the host.

To accomplish this, add a new boolean value to RDP surfaces, which
indicates, whether such local copy is necessary or not.
By default, the local copy is necessary.
As soon as an NVENC session is created, this boolean value is set to
TRUE.
When the RDP surface is resized or destroyed, this value is reset to
FALSE again.
Since all framebuffers are tracked, this won't cause any problems, when
resizing the associated stream.

This boolean value causes, that dma-bufs are no longer downloaded, when
it is set to TRUE.
The content of the dma-buf texture is just retrieved in a pixel buffer
object, which will be then mapped.
This will drastically reduce the frame latency due to the missing
download operation.
Also, when dma-bufs are not supported, the frame latency can be reduced
too, when memfds are used (which is usually the case, when dma-bufs are
not supported).
In this case, the mapped data does not need to be duplicated any more,
since the source stride is in almost all cases the same value as the
destination stride.
The EGL thread will then upload the mapped source content, instead of
uploading a copy of the source content.

The latency reduction depends on the framebuffer size and on the
system.
For FullHD frames, this is usually about 4ms.