rdp: Allow passing virtual monitor settings to pipewire-stream instance
When recording a virtual monitor, properties, such as the resolution,
are passed as pipewire-stream-properties.

Add a new parameter to grd_rdp_pipewire_stream_new() for a virtual
monitor.
This virtual monitor is just a struct with the monitor properties.

For remote assistance sessions, where the stream size is determined by
mutter, this virtual monitor will be NULL.

For headless sessions, this virtual monitor will be non-NULL and its
size is used, when connecting the pipewire stream.
mutter will then create a new virtual monitor with the passed size.