session: Also clear signal handler, when the session proxy is closed
When the remote desktop session is closed from the mutter side,
gnome-remote-desktop will clear the proxy and stop the session.
When this happens, the signal handler for the CAPS-lock and NUM-lock
signals are not set to 0, despite the remote desktop proxy already
been cleared.
This situation does not lead to a crash, but to unpleasant error
messages.

To fix this situation, clear the signal handlers before clearing the
remote desktop session proxy.
Usually this situation happens, when calling methods, that make mutter
close the remote desktop session.
One such situation is, when attempting to create and record a virtual
monitor in an X11 session.