vnc-pipewire-stream: Do not resize framebuffer when client is gone
When the VNC client is gone, the local rfb_client pointer is set to
NULL.
Since VNC sessions are not directly stopped, other GSources can run
before an idle GSource stops the session.
Such GSource can be the pipewire_source. If a VNC client closes the
connection after a stream is created, but before the framebuffer size
is known, gnome-remote-desktop crash, when resizing the framebuffer, as
rfb_client may already be NULL.

To fix this situation, check whether the client is gone, before
attempting to resize the framebuffer.

Fixes: https://errors.ubuntu.com/problem/ab880706e8a266568463cc4869ad9234e25a7d3b

https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/70