hwaccel-nvidia: Use a block dim of 16x16x1 for BGRX_TO_YUV420 kernel
Since H.264 works on 16x16 tiles, the NV12 buffer for the AVC encoder
is created in a way to include additional padding if the width or
height of the frame are not multiples of 16.
For the block dimension for the launch kernel call 32x8 is currently
used.
However, 16x16 would be the ideal block dimension, as in that case no
thread in a block would ever be idle.

So, change the block dimension to 16x16. The block size itself will
remain the same with this change (256).