rdp-gfx-surface: Don't forget to set the surface serial
Currently, the serial of a GFX surface is not set, making the serial
for each GFX surface effectively 0.
This is currently not a problem. But with virtual monitors and dynamic
resolution, this can lead to crashes in surface_serial_unref() in the
graphics pipeline, when not all frame acks are received, but the
virtual monitor has been recreated.
In that case the surface context is replaced and the old reference
count is lost, leading to an assert hit, when a frame ack is received
and the reference of the old surface serial is decremented.

Fix this issue by setting the surface serial, when creating a new GFX
surface.